tts-to-pubsub/requirements.txt
google-cloud
google-cloud-pubsub
functions-framework